Move the ModalResponse enum from the filesystem module to the appkit::app module.
It is now being used outside of filesystem, and this more closely mirrors
the location of this enum with AppKit.
