+++ src/filesystem/enums.rs
@@ -2,71 +2,6 @@
 
 use crate::foundation::{NSInteger, NSUInteger};
 
-/// Represents a modal response for macOS modal dialogs.
-#[derive(Copy, Clone, Debug)]
-pub enum ModalResponse {
-    /// The user hit the "Ok" button.
-    Ok,
-
-    /// Continue.
-    Continue,
-
-    /// Canceled.
-    Canceled,
-
-    /// Stopped.
-    Stopped,
-
-    /// Aborted.
-    Aborted,
-
-    /// The first button in the dialog was clicked.
-    FirstButtonReturned,
-
-    /// The second button in the dialog was clicked.
-    SecondButtonReturned,
-
-    /// The third button in the dialog was clicked.
-    ThirdButtonReturned
-}
-
-impl From<NSInteger> for ModalResponse {
-    fn from(i: NSInteger) -> Self {
-        match i {
-            1 => ModalResponse::Ok,
-            0 => ModalResponse::Canceled,
-            1000 => ModalResponse::FirstButtonReturned,
-            1001 => ModalResponse::SecondButtonReturned,
-            1002 => ModalResponse::ThirdButtonReturned,
-            -1000 => ModalResponse::Stopped,
-            -1001 => ModalResponse::Aborted,
-            -1002 => ModalResponse::Continue,
-
-            // @TODO: Definitely don't panic here, wtf was I thinking?
-            // Probably make this a ModalResponse::Unknown or something so a user can
-            // gracefully handle.
-            e => {
-                panic!("Unknown NSModalResponse sent back! {}", e);
-            }
-        }
-    }
-}
-
-impl Into<NSInteger> for ModalResponse {
-    fn into(self) -> NSInteger {
-        match self {
-            ModalResponse::Ok => 1,
-            ModalResponse::Canceled => 0,
-            ModalResponse::FirstButtonReturned => 1000,
-            ModalResponse::SecondButtonReturned => 1001,
-            ModalResponse::ThirdButtonReturned => 1002,
-            ModalResponse::Stopped => -1000,
-            ModalResponse::Aborted => -1001,
-            ModalResponse::Continue => -1002
-        }
-    }
-}
-
 /// Represents a type of search path used in file manager calls.
 #[derive(Copy, Clone, Debug)]
 pub enum SearchPathDomainMask {
